+++ src/com/suptrips/UsersDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface UsersDao  {
 
     Users addUser(Users users);
-    Users findUserById(int idbooster);
+    Users findUserById(long idbooster);
     List<Users> getAllUsers();
     void updateUsers(Users users);
     void removeUsers(Users users);
